+++ plugins/helper/EdkKillboardHelper.php
@@ -80,9 +80,9 @@ class EdkKillboardHelper {
 			 * Overwrite Victim Image if its a Citadel
 			 */
 			if(\in_array($kill->shp_name, EveOnline\Plugins\Killboard::getStructureNames())) {
-				$kill->victimImage = $this->getStructureImage($kill->shp_id);
+				$kill->victimImage = $this->getStructureImage($kill->shp_id, 256);
 			} else {
-				$kill->victimImage = $this->getVictimImage($kill->plt_name, $kill->shp_id);
+				$kill->victimImage = $this->getVictimImage($kill->plt_name, $kill->shp_id, 256);
 			} // END if(\in_array($kill->shp_name, $citadelNames))
 		} // END foreach($resultLastKills as &$kill)